+++ src/providers/types.ts
@@ -1,6 +1,5 @@
 import type { BaseChatModel } from '@langchain/core/language_models/chat_models';
 import type { EmbeddingsInterface } from '@langchain/core/embeddings';
-import type { LangChainOrchestrationModuleConfig } from '@sap-ai-sdk/langchain';
 
 export type ModelOptions = Record<string, unknown>;
 
@@ -31,13 +30,12 @@ export interface BuiltInProviderOptions {
     defaultEmbeddingModel?: string;
 }
 
-export interface SapAICoreModelEntry extends BuiltInProviderModelMapEntry {
-    config?: LangChainOrchestrationModuleConfig;
-    langchainOptions?: ModelOptions;
-}
+export interface SapAICoreModelEntry extends BuiltInProviderModelMapEntry { }
 
 export interface SapAICoreProviderOptions {
     chatModels?: Record<string, SapAICoreModelEntry>;
+    embeddingModels?: Record<string, SapAICoreModelEntry>;
     defaultChatModel?: string;
+    defaultEmbeddingModel?: string;
 }